scaffoldで追加したcustomerモデルのファイル・ルーティングをadmins/の直下へ移動＆各ファイル内のpathなど影響箇所を修正した
