+++ app/views/admins/customers/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: customer) do |form| %>
+<%= form_with model: customer, url: url do |form| %>
   <% if customer.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(customer.errors.count, "error") %> prohibited this customer from being saved:</h2>
