+++ app/views/admins/customers/edit.html.erb
@@ -1,0 +1,6 @@
+<h1>Editing Customer</h1>
+
+<%= render 'form', customer: @customer, url: admins_customer_path(@customer) %>
+
+<%= link_to 'Show', admins_customer_path(@customer) %> |
+<%= link_to 'Back', admins_customers_path %>
