+++ app/views/admins/customers/index.html.erb
@@ -16,9 +16,9 @@
       <tr>
         <td><%= customer.name %></td>
         <td><%= customer.email %></td>
-        <td><%= link_to 'Show', customer %></td>
-        <td><%= link_to 'Edit', edit_customer_path(customer) %></td>
-        <td><%= link_to 'Destroy', customer, method: :delete, data: { confirm: 'Are you sure?' } %></td>
+        <td><%= link_to 'Show', admins_customer_path(customer) %></td>
+        <td><%= link_to 'Edit', edit_admins_customer_path(customer) %></td>
+        <td><%= link_to 'Destroy', admins_customer_path(customer), method: :delete, data: { confirm: 'Are you sure?' } %></td>
       </tr>
     <% end %>
   </tbody>
@@ -26,4 +26,4 @@
 
 <br>
 
-<%= link_to 'New Customer', new_customer_path %>
+<%= link_to 'New Customer', new_admins_customer_path %>
