+++ app/views/admins/customers/new.html.erb
@@ -1,0 +1,5 @@
+<h1>New Customer</h1>
+
+<%= render 'form', customer: @customer, url: admins_customers_path(@customer) %>
+
+<%= link_to 'Back', admins_customers_path %>
