+++ app/views/admins/customers/show.html.erb
@@ -10,5 +10,5 @@
   <%= @customer.email %>
 </p>
 
-<%= link_to 'Edit', edit_customer_path(@customer) %> |
-<%= link_to 'Back', customers_path %>
+<%= link_to 'Edit', edit_admins_customer_path(@customer) %> |
+<%= link_to 'Back', admins_customers_path %>
