--- app/views/customers/edit.html.erb
@@ -1,6 +1,0 @@
-<h1>Editing Customer</h1>
-
-<%= render 'form', customer: @customer %>
-
-<%= link_to 'Show', @customer %> |
-<%= link_to 'Back', customers_path %>
