--- app/views/customers/new.html.erb
@@ -1,5 +1,0 @@
-<h1>New Customer</h1>
-
-<%= render 'form', customer: @customer %>
-
-<%= link_to 'Back', customers_path %>
